prod-requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
daphne==4.1.2
redis==6.0.0
Django==4.2.10
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
python-decouple==3.8
python-dateutil==2.9.0.post0
PyJWT==2.10.1
requests==2.32.3
stripe==11.6.0
psycopg2-binary==2.9.10
pillow==11.1.0
pytz==2025.1
tzdata==2024.2
sqlparse==0.5.3
numpy==1.26.4
pandas==2.2.2
typing_extensions==4.12.2
